Move tests around
also fix wrong package names
