+++ kotlinx-coroutines-core/jvm/test/scheduling/RunBlockingCoroutineSchedulerLivenessStressTest.kt
@@ -1,13 +1,46 @@
-package scheduling
+package kotlinx.coroutines.scheduling
 
 import kotlinx.coroutines.*
-import kotlinx.coroutines.scheduling.*
 import kotlinx.coroutines.testing.*
+import org.junit.*
+import org.junit.runner.*
+import org.junit.runners.*
 import java.util.*
 import java.util.concurrent.*
 
-open class RunBlockingCoroutineSchedulerLivenessTestBase : SchedulerTestBase() {
-    protected fun testSchedulerLiveness(targetDispatcher: CoroutineDispatcher, yieldMask: Int = 0b1111): Unit = runBlocking {
+@RunWith(Parameterized::class)
+class RunBlockingCoroutineSchedulerLivenessStressTest(private val yieldMask: Int) : SchedulerTestBase() {
+    init {
+        corePoolSize = 1
+    }
+
+    companion object {
+        @JvmStatic
+        @Parameterized.Parameters
+        fun data(): Array<Array<Any?>> {
+            return Array(16 * stressTestMultiplierSqrt) { arrayOf(it) }
+        }
+    }
+
+    @Test
+    fun testLivenessOfDefaultDispatcher(): Unit = testSchedulerLiveness(dispatcher, yieldMask)
+
+    @Test
+    fun testLivenessOfIoDispatcher(): Unit = testSchedulerLiveness(blockingDispatcher(1), yieldMask)
+
+    @Test
+    fun testLivenessOfLimitedDispatcherOnTopOfDefaultDispatcher() =
+        testSchedulerLiveness(dispatcher.limitedParallelism(1), yieldMask)
+
+    @Test
+    fun testLivenessOfLimitedDispatcherOnTopOfIoDispatcher() = testSchedulerLiveness(
+        // Important: inner limitedDispatcher will be on top of this LimitedDispatcher, so there are two Workers from
+        // two different LimitedDispatchers that must coordinate their permits, not just one.
+        // In other words, LimitedDispatcher's Worker should also respect BlockingDispatchAware on its inner tasks
+        blockingDispatcher.value.limitedParallelism(1), yieldMask
+    )
+
+    private fun testSchedulerLiveness(targetDispatcher: CoroutineDispatcher, yieldMask: Int = 0b1111): Unit = runBlocking {
         val oldRunBlockings = LinkedList<Job>()
         var maxOldRunBlockings = 0
         var busyWaits = 0
